TASK05: Add helloworld /proc File
Here a simple example showing how to use a /proc file.
This is the HelloWorld for the /proc filesystem.
There are three parts: create the file /proc/helloworld
in the function init_module, return a value (and a buffer)
when the file /proc/helloworld is read in the callback
function procfile_read, and delete the file /proc/helloworld
in the function cleanup_module.

The /proc/helloworld is created when the module is loaded
with the function proc_create.
The return value is a struct proc_dir_entry,
and it will be used to configure the file /proc/helloworld
(for example, the owner of this file).
A null return value means that the creation has failed.

Every time the file /proc/helloworld is read,
the function procfile_read is called. Two parameters of this function
are very important: the buffer (the second parameter)
and the offset (the fourth one).
The content of the buffer will be returned to the application
which read it (for example the cat command).
The offset is the current position in the file.
If the return value of the function is not null,
then this function is called again.
So be careful with this function, if it never returns zero,
the read function is called endlessly.

$ cat /proc/helloworld
HelloWorld!

Signed-off-by: Андрій <andrii@sphinx.netis.cc>
Signed-off-by: Andrii Anoshyn <anoshyn.andrii@gmail.com>