TASK06: Add ioctl (clear, set, get buffer size)
Added an ioctl driver for a character device that implements calls
to clear the buffer, get the size of the buffer and set the new size
of the buffer.
In the user space - use main.c program with commands:
- ./main clr - for clear buffer;
- ./main get - for get size of the buffer;
- ./main set <new size> - for set new size of the buffer,
successful execution - return the new buffer size.
chrdev.ko and main - compiled on the Raspberry Pi Zero HW
(Linux 5.15.84+, arm6l).
Added the third part of the task to README file.
File chrdev.c, main.c and cdev_ioctl.h is checked with checkpatch.pl.

Signed-off-by: Sergiy Us <sergiy.us@gmail.com>