fp-ultrasound: Change - Makefile, platform_driver.c; Move - *.*; Add - dt_ultrasound.dtbo, ultrasound_mod.ko
Filled Makefile to build module of kernel and *.dtbo files
Change platform_driver.c to module, when property is not present, return current value of error.
Builded dt_ultrasound.dtbo as device tree blob overlay to download it in kernel.
Builded ultrasound_mod.ko as module of kernel to download it in kernel.
Moved all files to a separate folder so that when merging branches, they are not mixed up

Signed-off-by: Roman Muravka <roman.muravka@nure.ua>